pyproject.toml
[project]
name = "htr-essays"
version = "0.1.0"
description = "HTR Pipeline for Swedish Student Essays using TrOCR"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.35.0",
    "Pillow>=10.0.0",
    "pdf2image>=1.16.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "tensorboard>=2.14.0",
    "jiwer>=3.0.0",
    "opencv-python>=4.8.0",
    "accelerate>=0.24.0",
    "datasets>=2.14.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
]

[tool.poetry]
packages = [{include = "htr_essays", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
